chore: Move generic postgres stuff out of player repository
Will use this for the upcoming username repository
