+++ internal/adapters/database/cloudsql.go
@@ -1,4 +1,4 @@
-package playerrepository
+package database
 
 import "fmt"
