+++ internal/adapters/database/migrator.go
@@ -1,4 +1,4 @@
-package playerrepository
+package database
 
 import (
 	"context"
