+++ internal/adapters/database/postgres_test.go
@@ -1,4 +1,4 @@
-package playerrepository
+package database
 
 import (
 	"fmt"
